requirements.txt
Pillow==11.3.0
fpdf2==2.8.3
